feat(core): integrate turbo for efficient development
