+++ package.json
@@ -24,7 +24,10 @@
     "create"
   ],
   "scripts": {
-    "dev": "pnpm -r dev",
+    "dev:core": "turbo dev --filter=./packages/core",
+    "dev:info": "turbo dev --filter=./packages/info",
+    "dev:shared": "turbo dev --filter=./packages/shared",
+    "dev": "turbo dev",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "typecheck": "tsc --noEmit"
@@ -39,6 +42,7 @@
     "simple-git-hooks": "^2.11.1",
     "terser": "^5.31.6",
     "tsup": "^8.2.4",
+    "turbo": "^2.3.3",
     "typescript": "^5.5.4"
   },
   "simple-git-hooks": {
