+++ packages/core/package.json
@@ -40,6 +40,7 @@
   },
   "devDependencies": {
     "@clack/prompts": "^0.9.0",
+    "@create-uni/shared": "workspace:*",
     "@types/cross-spawn": "^6.0.6",
     "@types/ejs": "^3.1.5",
     "@types/minimist": "^1.2.5",
