+++ packages/info/package.json
@@ -39,11 +39,11 @@
   },
   "devDependencies": {
     "@clack/prompts": "^0.9.0",
+    "@create-uni/shared": "workspace:*",
     "@types/envinfo": "^7.8.4",
     "@types/minimist": "^1.2.5",
     "cross-env": "^7.0.3",
     "envinfo": "^7.14.0",
-    "json5": "^2.2.3",
     "kolorist": "^1.8.0",
     "path-exists": "^5.0.0"
   }
