+++ packages/info/src/constants/index.ts
@@ -1,0 +1,7 @@
+export const uniDependenciesMap = {
+  '@uni-helper/uni-use': ['@vueuse/core'],
+  '@uni-helper/axios-adapter': ['axios'],
+  '@uni-helper/unocss-preset-uni': ['unocss'],
+  '@uni-helper/eslint-config': ['eslint'],
+  '@uni-helper/vite-plugin-uni-tailwind': ['tailwindcss'],
+} as Record<string, string[]>
