+++ packages/shared/tsup.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from 'tsup'
+
+export default defineConfig({
+  entry: ['src/index.ts'],
+  format: 'esm',
+  platform: 'node',
+  target: 'node14',
+  clean: true,
+  dts: true,
+})
